# Optimize dispatcher by using positional only arguments