--- a/doc/changes/2135.misc
+++ b/doc/changes/2135.misc
@@ -1,0 +1,1 @@
+Optimize dispatcher by dispatching on positional only args.