--- a/qutip/core/data/adjoint.pyx
+++ b/qutip/core/data/adjoint.pyx
@@ -114,7 +114,7 @@ import inspect as _inspect
 
 adjoint = _Dispatcher(
     _inspect.Signature([
-        _inspect.Parameter('matrix', _inspect.Parameter.POSITIONAL_OR_KEYWORD),
+        _inspect.Parameter('matrix', _inspect.Parameter.POSITIONAL_ONLY),
     ]),
     name='adjoint',
     module=__name__,
@@ -129,7 +129,7 @@ adjoint.add_specialisations([
 
 transpose = _Dispatcher(
     _inspect.Signature([
-        _inspect.Parameter('matrix', _inspect.Parameter.POSITIONAL_OR_KEYWORD),
+        _inspect.Parameter('matrix', _inspect.Parameter.POSITIONAL_ONLY),
     ]),
     name='transpose',
     module=__name__,
@@ -144,7 +144,7 @@ transpose.add_specialisations([
 
 conj = _Dispatcher(
     _inspect.Signature([
-        _inspect.Parameter('matrix', _inspect.Parameter.POSITIONAL_OR_KEYWORD),
+        _inspect.Parameter('matrix', _inspect.Parameter.POSITIONAL_ONLY),
     ]),
     name='conj',
     module=__name__,