… qutip/core/data/inner.pyx
… qutip/core/data/kron.pyx
… qutip/core/data/linalg.py
… qutip/core/data/matmul.pyx
… qutip/core/data/mul.pyx
… qutip/core/data/norm.pyx
… qutip/core/data/permute.pyx
… qutip/core/data/pow.pyx
… qutip/core/data/project.pyx
… qutip/core/data/properties.pyx
… qutip/core/data/ptrace.pyx
… qutip/core/data/reshape.pyx
… qutip/core/data/solve.py
… qutip/core/data/tidyup.pyx
… qutip/core/data/trace.pyx