**Description**
I optimized the dispatched by only supporting positional arguments for matrix to dispatch on:

The dispatcher was supporting any signature for dispatched functions.
Since each dispatchers is an instance of the `Dispatcher` class and cython does not support patching methods, it had to find the inputs matrix to convert from `__call__(*args, **kwargs)` and re-implemented the `_bind` method to do so.
But all our dispatched functions have the matrix as the first positional argument, by embracing this we get a nice speed up for small matrices.

Also when the specialization existed, it would not call it directly, but call  a `_constructed_specialisation` that called ti. Adding another unneeded layer to each calls. This was changed so it would call the function directly.

For a 2x2 matrix:
|                  | Before | After |
|------------------|--------|-------|
| data.add_dense   | 214ns  | 210ns |
| data.add (dense) | 774ns  | 388ns |
| data.add[Dense, Dense] | 541ns  | 182ns |
| data.add_csr     | 429ns  | 447ns |
| data.add (csr)   | 986ns  | 649ns |
| data.add[CSR. CSR]  | 744ns  | 401ns |
| numpy | 473ns | - |

Sadly it will not improve the benchmarks that much since the `Qobj` operations overhead is quite large:
|                  | Before | After |
|------------------|--------|-------|
| Qboj + Qobj (dense)   | 2430ns  | 2070ns |
| Qboj + Qobj (csr)   | 3070ns  | 2330ns |